requirements.txt
anyio==4.7.0
aiogram-3.16.0
certifi==2024.8.30
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
python-telegram-bot==21.9
sniffio==1.3.1
typing_extensions==4.12.2
